Clarify Console info and RT correction commands
