+++ tests/MSDIAL5/MsdialCoreTestApp/Program.cs
@@ -183,8 +183,9 @@ class Program {
         MainProcess.SetImmsCommand(root);
         MainProcess.SetMsnCommand(root);
         MainProcess.SetEicCommand(root);
+        MainProcess.SetRtCorrectionCommand(root);
         MainProcess.SetImageGenerationCommand(root);
-        MainProcess.SetCapabilitiesCommand(root);
+        MainProcess.SetInfoCommand(root);
         var parseResult = root.Parse(args);
         return parseResult.InvokeAsync();
     }
